feat: swtich to units API (from file API)
The weblate API to get translation files performs a commit in the Git
repository which is usually not intended when fetching the translation
strings. That's why we change to the translation units API.

The units API also provides the possibility to only load strings that were
changed since a given timestamp. This helps in reducing the load to the
server when filling the cache.

Closes #38