fix: loading translations via changed date
When using the units API and getting the translations via
`"changed:>=YYYY-MM-DD"` you will not get all translations wanted.
Translations which were only once touched get only an added date.
So, those translations will not be fetched.

This changes query for the `changed` and `added` fields to fix this.

Fixes #43